Rename & Cleanup of clientslist container
