+++ src/containers/Clients/List/tests/index.test.tsx
@@ -3,10 +3,10 @@ import { mount, shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import * as React from 'react';
 import { Provider } from 'react-redux';
+import ConnectedClientsList, { ClientList } from '..';
 import store from '../../../../store';
 import reducer, { fetchClients, reducerName } from '../../../../store/ducks/clients';
 import * as fixtures from '../../../../store/ducks/tests/fixtures';
-import ConnectedClientsList, { ClientList } from '../clientList';
 
 reducerRegistry.register(reducerName, reducer);
